# api/docs/admin/admin_swagger.yaml
basePath: /api
definitions:
...
  admin.TagForm:
    properties:
      collection_id:
        type: integer
      color:
        type: integer
      id:
        type: integer
      name:
        type: string
      user_id:
        type: integer
    required:
    - color
    - name
    type: object
...
  model.AddressBookCollection:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - name
    type: object
  model.AddressBookCollectionList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.AddressBookCollection'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  model.AddressBookCollectionRule:
    properties:
      collection_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      rule:
        description: '0: 无 1: 读 2: 读写  3: 完全控制'
        maximum: 3
        minimum: 1
        type: integer
      to_id:
        type: integer
      type:
        description: '1: 个人 2: 群组'
        maximum: 2
        minimum: 1
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - collection_id
    - rule
    - to_id
    - type
    type: object
...
  model.Tag:
    properties:
      collection:
        $ref: '#/definitions/model.AddressBookCollection'
      collection_id:
        type: integer
      color:
        description: color 是flutter的颜色值,从0x00000000 到 0xFFFFFFFF; 前两位表示透明度，后面6位表示颜色,
          可以转成rgba
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  model.TagList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 接口
  title: 管理系统API
  version: "1.0"
paths:
...
  /admin/my/address_book_collection_rule/create:
    post:
      consumes:
      - application/json
      description: 创建地址簿规则
      parameters:
      - description: 地址簿规则信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollectionRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollection'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 创建地址簿规则
      tags:
      - 我的地址簿规则
  /admin/my/address_book_collection_rule/delete:
    post:
      consumes:
      - application/json
      description: 地址簿规则删除
      parameters:
      - description: 地址簿规则信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollectionRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿规则删除
      tags:
      - 我的地址簿规则
  /admin/my/address_book_collection_rule/list:
    get:
      consumes:
      - application/json
      description: 地址簿规则列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页大小
        in: query
        name: page_size
        type: integer
      - description: 是否是我的
        in: query
        name: is_my
        type: integer
      - description: 用户id
        in: query
        name: user_id
        type: integer
      - description: 地址簿集合id
        in: query
        name: collection_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollectionList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿规则列表
      tags:
      - 我的地址簿规则
  /admin/my/address_book_collection_rule/update:
    post:
      consumes:
      - application/json
      description: 地址簿规则编辑
      parameters:
      - description: 地址簿规则信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollectionRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollection'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿规则编辑
      tags:
      - 我的地址簿规则
...
  /admin/my/tag/create:
    post:
      consumes:
      - application/json
      description: 创建标签
      parameters:
      - description: 标签信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.TagForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Tag'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 创建标签
      tags:
      - 我的标签
  /admin/my/tag/delete:
    post:
      consumes:
      - application/json
      description: 标签删除
      parameters:
      - description: 标签信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.TagForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 标签删除
      tags:
      - 标签
  /admin/my/tag/list:
    get:
      consumes:
      - application/json
      description: 标签列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页大小
        in: query
        name: page_size
        type: integer
      - description: 是否是我的
        in: query
        name: is_my
        type: integer
      - description: 用户id
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.TagList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 标签列表
      tags:
      - 我的标签
  /admin/my/tag/update:
    post:
      consumes:
      - application/json
      description: 标签编辑
      parameters:
      - description: 标签信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.TagForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Tag'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 标签编辑
      tags:
      - 我的标签
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"